fix: rename selection file to prevent test discovery
Rename test_selection.py to selection.py to prevent pytest considering it as a test
