+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
-norecursedirs = pytest_rts/tests/helper_project pytest_rts/utils
+# Prevent helper project tests being collected as tests for pytest-rts
+norecursedirs = pytest_rts/tests/helper_project
